Import prebid as a module
